requirements.txt
# ============================================
# EduReflex 多智能体协作学习系统 - 依赖清单
# ============================================

# 核心框架：LangGraph 状态机（让 pip 自动解决版本依赖）
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# 大模型接口（支持 DeepSeek 和 OpenAI）
openai>=1.0.0
langchain-openai>=0.2.0

# 后端框架：FastAPI（异步高性能）
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# 持久化与缓存
redis>=5.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 搜索工具（免费方案）
# 注意：duckduckgo-search 已改名为 ddgs
ddgs>=0.1.0

# HTTP 请求（调用 RAG 接口）
httpx>=0.27.0
aiohttp>=3.9.0

# 工具类
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日志与监控
loguru>=0.7.0

# 测试工具（可选）
# pytest==8.3.3
# pytest-asyncio==0.24.0
# locust==2.32.2  # 并发压测

